backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
google-auth==2.23.3
google-api-python-client==2.100.0
requests==2.31.0
python-multipart==0.0.6
Pillow>=10.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
